RT#102626: proper fix for t/parent-pmc.t
A failing case of t/parent-pmc.t was remaining: perl < 5.14 with
PERL_DISABLE_PMC in build options.
This is now fixed. We check for PERL_DISABLE_PMC using Config::Perl::V.
This module is added as a 'recommends' test dependency in Makefile.PL. If the
dependency is missing or if PERL_DISABLE_PMC is set we skip the test.
On smokers ($ENV{SMOKE_TESTING}=1 and $] < 5.14) we make the dependency
mandatory.