t/parent.t - Harden test against possible wording change
In https://github.com/Perl/perl5/pull/20547 we will likely change this
error message to be more appropriate. This patch hardens the tests so
that whatever wording is chosen will be accepted without error, and so
the tests can run on whatever perl is used.

Eg, in the future perl may say "@INC entries checked:" instead of
"@INC contains:", as we will start reporting what was actually done
instead of the state of @INC at the end of the require. The two may
not be even remotely the same. This patch ensures that whatever wording
is used after the @INC and before the ":" we will pass test.